[ci][bisect/0] pass Test object into bisector
