+++ ci/ray_ci/bisect/test_bisector.py
@@ -4,6 +4,7 @@ from unittest import mock
 
 from ci.ray_ci.bisect.bisector import Bisector
 from ci.ray_ci.bisect.macos_validator import MacOSValidator
+from ray_release.test import Test
 
 
 @mock.patch("ci.ray_ci.bisect.bisector.Bisector._checkout_and_validate")
@@ -16,10 +17,10 @@ def test_run(mock_get_revision_lists, mock_checkout_and_validate):
     mock_get_revision_lists.return_value = ["1", "2", "3", "4", "5"]
 
     # Test case 1: P P P F F
-    assert Bisector("test", "1", "5", MacOSValidator()).run() == "3"
+    assert Bisector(Test(), "1", "5", MacOSValidator()).run() == "3"
 
     # Test case 2: P F
-    assert Bisector("test", "3", "4", MacOSValidator()).run() == "3"
+    assert Bisector(Test(), "3", "4", MacOSValidator()).run() == "3"
 
 
 if __name__ == "__main__":
