+++ ci/ray_ci/bisect/validator.py
@@ -1,9 +1,11 @@
 import abc
 
+from ray_release.test import Test
+
 
 class Validator(abc.ABC):
     @abc.abstractmethod
-    def run(self, test: str) -> bool:
+    def run(self, test: Test) -> bool:
         """
         Validate whether the test is passing or failing on the given revision
         """
